requirements.txt
b3j0f.conf>=0.3.13
b3j0f.middleware>=0.0.3
jsonpatch>=1.13
jsonschema>=2.5.1
